Make FlashingFire more accurate
Previously FlashingFire would just compare damage and health to determine if it should add back.
This would not work if a card had a damage cap (as well as many other potential damage prevention mechanisms).
Now it just checks the target's location after dealing the damage, which is much more reliable.